// src/pages/WechatManage/Material/data.d.ts
export interface TableListItem {
    media_id: string,
    title: string,
    author: string,
    digest: string,
    content_source_url: string,
    content: string,
    url: string,
    thumb_url: string,
    need_open_comment: string,
    only_fans_can_comment: string,
    create_time: Date,
    update_time: Date
}